+++ ad-service/ad-search/src/main/java/com/glw/ad/index/unit/AdUnitObject.java
@@ -18,9 +18,9 @@ public class AdUnitObject {
 
     private Long unitId;
 
-    private Integer unitStatus;
+    private Byte unitStatus;
 
-    private Integer positionType;
+    private Byte positionType;
 
     private Long planId;
